Fixing name change revert if layout had new changes right after (#441)
**User-Facing Changes**
Layouts will no longer revert the name change if an user applies changes
to it, like changing panels or readjusting panels height.

**Description**
Basically the error was occurring because, despite the layout state
being correctly updated when the name was changed, the layout in
question was still selected and that selection would keep the older
state,, so when an user tried to make any change to the layout, that
change would be done with the `layoutState` of what was selected, and
therefore not the updated version.
My fix to this issue was simply refresh the selected layout by grabbing
its id after the name was updated and therefore updating its
`layoutState`.

**Checklist**

- [X] The web version was tested and it is running ok
- [X] The desktop version was tested and it is running ok
- [X] This change is covered by unit tests
- [X] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated